move eas update section into its own file
